fix: move data fetching from useHeatmapSelection
